Separate algebra tests.
