+++ tests/algebra/tuples.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   operations.c                                       :+:      :+:    :+:   */
+/*   tuples.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/06 16:59:18 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/12/28 16:03:45 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/12/28 16:11:06 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,6 @@
 t_t3d expected;
 t_t3d result;
 t_t3d _tuple;
-
-void assert_tuple_eq(t_t3d _expected, t_t3d _result)
-{
-	mu_assert_double_eq(_expected.x, _result.x);
-	mu_assert_double_eq(_expected.y, _result.y);
-	mu_assert_double_eq(_expected.z, _result.z);
-	mu_assert_double_eq(_expected.type, _result.type);
-}
 
 void test_setup(void)
 {
@@ -39,22 +31,6 @@ MU_TEST(tuple_tst)
 
 	result = tuple(4.3, -4.2, 3.1, POINT_TYPE);
 	expected = (t_t3d){4.3, -4.2, 3.1, POINT_TYPE};
-	assert_tuple_eq(expected, result);
-}
-
-MU_TEST(vector_tst)
-{
-	result = vector(43.0, -341.94, 11.421);
-	expected = (t_v3d){43.0, -341.94, 11.421, VECTOR_TYPE};
-
-	assert_tuple_eq(expected, result);
-}
-
-MU_TEST(point_tst)
-{
-	result = point(43.0, -341.94, 11.421);
-	expected = (t_p3d){43.0, -341.94, 11.421, POINT_TYPE};
-
 	assert_tuple_eq(expected, result);
 }
 
@@ -149,32 +125,11 @@ MU_TEST(scalar_div_tst)
 	assert_tuple_eq(expected, result);
 }
 
-MU_TEST(magnitude_tst)
-{
-	_tuple = vector(1, 0, 0);
-	mu_assert_double_eq(1, magnitude(_tuple));
-
-	_tuple = vector(0, 1, 0);
-	mu_assert_double_eq(1, magnitude(_tuple));
-
-	_tuple = vector(0, 0, 1);
-	mu_assert_double_eq(1, magnitude(_tuple));
-
-	_tuple = vector(1, 2, 3);
-	mu_assert_double_eq(sqrt(14), magnitude(_tuple));
-
-	_tuple = vector(-1, -2, -3);
-	mu_assert_double_eq(sqrt(14), magnitude(_tuple));
-}
-
-
 MU_TEST_SUITE(vectors_suite)
 {
 	MU_SUITE_CONFIGURE(&test_setup, &test_teardown);
 
 	MU_RUN_TEST(tuple_tst);
-	MU_RUN_TEST(vector_tst);
-	MU_RUN_TEST(point_tst);
 
 	MU_RUN_TEST(add_tst);
 	MU_RUN_TEST(sub_tst);
@@ -182,8 +137,6 @@ MU_TEST_SUITE(vectors_suite)
 
 	MU_RUN_TEST(scalar_times_tst);
 	MU_RUN_TEST(scalar_div_tst);
-
-	MU_RUN_TEST(magnitude_tst);
 }
 
 MU_MAIN
